refactor(RequestDto): 요청 DTO들의 package를 재이동시켰습니다
Request관련 DTO -> presentation 패키지로 재이동
- application 계층은 presentation과 domain을 연결해주는 계층입니다. 따라서, 해당 계층은 요청 dto를 presentation으로 부터 전달 받아, domain 계층과 상호작용 후, 응답 dto를 내려주는 역할을 합니다.
- 따라서, 요청 dto와 응답 dto는 presentation 계층이 더 어울린다는 판단이 들었습니다. 만약 데이터 전송 객체가 필요하다면, application 계층에는 dao와 같은 객체가 더 어울릴지도 모른다는 생각이 드네요.
