--- src/main/java/com/realworld/study/post/presentation/dto/PostDeleteResponse.java
@@ -1,4 +1,0 @@
-package com.realworld.study.post.presentation.dto;
-
-public record PostDeleteResponse(boolean isDeleted) {
-}
