+++ src/main/java/com/realworld/study/post/presentation/dto/request/PostCreateRequest.java
@@ -1,4 +1,4 @@
-package com.realworld.study.post.application.dto;
+package com.realworld.study.post.presentation.dto.request;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -6,11 +6,11 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class PostUpdateRequest {
+public class PostCreateRequest {
     private String title;
     private String contents;
 
-    public PostUpdateRequest(String title, String contents) {
+    public PostCreateRequest(String title, String contents) {
         this.title = title;
         this.contents = contents;
     }
