+++ src/main/java/com/realworld/study/post/presentation/dto/request/PostUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.realworld.study.post.application.dto;
+package com.realworld.study.post.presentation.dto.request;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -6,11 +6,11 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class PostCreateRequest {
+public class PostUpdateRequest {
     private String title;
     private String contents;
 
-    public PostCreateRequest(String title, String contents) {
+    public PostUpdateRequest(String title, String contents) {
         this.title = title;
         this.contents = contents;
     }
