+++ src/main/java/com/realworld/study/post/presentation/dto/response/PostDeleteResponse.java
@@ -1,0 +1,4 @@
+package com.realworld.study.post.presentation.dto.response;
+
+public record PostDeleteResponse(boolean isDeleted) {
+}
